+++ MatrixSDKTests/MXRoomStateTests.m
@@ -820,10 +820,10 @@
                     
                     XCTAssertNotNil(newRoom);
                     
-                    XCTAssertEqual(newRoom.state.membership, MXMembershipInvite);
+                    XCTAssertEqual(newRoom.summary.membership, MXMembershipInvite);
                     
                     // The room has 2 members (Alice & Bob)
-                    XCTAssertEqual(newRoom.state.membersCount.members, 2);
+                    XCTAssertEqual(newRoom.summary.membersCount.members, 2);
 
                     MXRoomMember *alice = [newRoom.state.members memberWithUserId:aliceRestClient.credentials.userId];
                     XCTAssertNotNil(alice);
@@ -868,12 +868,12 @@
                         
                         XCTAssertNotNil(newRoom);
 
-                        if (newRoom.state.membership != MXMembershipUnknown)
+                        if (newRoom.summary.membership != MXMembershipUnknown)
                         {
-                            XCTAssertEqual(newRoom.state.membership, MXMembershipInvite);
+                            XCTAssertEqual(newRoom.summary.membership, MXMembershipInvite);
 
                             // The room has 2 members (Alice & Bob)
-                            XCTAssertEqual(newRoom.state.membersCount.members, 2);
+                            XCTAssertEqual(newRoom.summary.membersCount.members, 2);
 
                             MXRoomMember *alice = [newRoom.state.members memberWithUserId:aliceRestClient.credentials.userId];
                             XCTAssertNotNil(alice);
@@ -953,10 +953,10 @@
                         
                         // Now, we must have more information about the room
                         // Check its new state
-                        XCTAssertEqual(newRoom.state.membersCount.members, 2);
+                        XCTAssertEqual(newRoom.summary.membersCount.members, 2);
                         XCTAssert([newRoom.state.topic isEqualToString:@"We test room invitation here"], @"Wrong topic. Found: %@", newRoom.state.topic);
                         
-                        XCTAssertEqual(newRoom.state.membership, MXMembershipJoin);
+                        XCTAssertEqual(newRoom.summary.membership, MXMembershipJoin);
 
                         XCTAssertNotNil(newRoom.summary.lastMessageEventId);
                         XCTAssertNotNil(newRoom.summary.lastMessageEvent);
@@ -1000,7 +1000,7 @@
                     
                     [mxSession joinRoom:roomId success:^(MXRoom *room) {
                         
-                        XCTAssert([room.state.roomId isEqualToString:roomId]);
+                        XCTAssert([room.roomId isEqualToString:roomId]);
                         
                         MXRoom *newRoom = [mxSession roomWithRoomId:roomId];
                         XCTAssert(newRoom, @"The room must be known now by the user");
@@ -1008,10 +1008,10 @@
                         // Now, we must have more information about the room
                         // Check its new state
                         XCTAssertEqual(newRoom.state.isJoinRulePublic, YES);
-                        XCTAssertEqual(newRoom.state.membersCount.members, 2);
+                        XCTAssertEqual(newRoom.summary.membersCount.members, 2);
                         XCTAssert([newRoom.state.topic isEqualToString:@"We test room invitation here"], @"Wrong topic. Found: %@", newRoom.state.topic);
                         
-                        XCTAssertEqual(newRoom.state.membership, MXMembershipJoin);
+                        XCTAssertEqual(newRoom.summary.membership, MXMembershipJoin);
                         XCTAssertNotNil(newRoom.summary.lastMessageEvent);
                         XCTAssertEqual(newRoom.summary.lastMessageEvent.eventType, MXEventTypeRoomMember, @"The last should be a m.room.member event indicating Alice joining the room");
                         
@@ -1116,7 +1116,7 @@
                     MXRoom *room = [mxSession roomWithRoomId:event.roomId];
 
                     XCTAssert(room);
-                    XCTAssertEqual(room.state.membersCount.members, 2, @"If this count is wrong, the room state is invalid");
+                    XCTAssertEqual(room.summary.membersCount.members, 2, @"If this count is wrong, the room state is invalid");
 
                     [expectation fulfill];
                 }
@@ -1165,7 +1165,7 @@
                 MXRoom *room = [mxSession roomWithRoomId:newRoomId];
                 XCTAssertNotNil(room);
                 
-                BOOL isSync = (room.state.membership != MXMembershipInvite && room.state.membership != MXMembershipUnknown);
+                BOOL isSync = (room.summary.membership != MXMembershipInvite && room.summary.membership != MXMembershipUnknown);
                 XCTAssertFalse(isSync, @"The room is not yet sync'ed");
 
                 [[NSNotificationCenter defaultCenter] removeObserver:newRoomObserver];
@@ -1178,9 +1178,9 @@
 
                 MXRoom *room = note.object;
 
-                XCTAssertEqualObjects(newRoomId, room.state.roomId);
+                XCTAssertEqualObjects(newRoomId, room.roomId);
                 
-                BOOL isSync = (room.state.membership != MXMembershipInvite && room.state.membership != MXMembershipUnknown);
+                BOOL isSync = (room.summary.membership != MXMembershipInvite && room.summary.membership != MXMembershipUnknown);
                 XCTAssert(isSync, @"The room must be sync'ed now");
 
                 [[NSNotificationCenter defaultCenter] removeObserver:initialSyncObserver];