+++ MatrixSDKTests/MXRoomTests.m
@@ -172,12 +172,12 @@
         
         mxSession = mxSession2;
         
-        NSString *roomId = room.state.roomId;
+        NSString *roomId = room.roomId;
 
         __block MXMembership lastKnownMembership = MXMembershipUnknown;
         [room.liveTimeline listenToEventsOfTypes:@[kMXEventTypeStringRoomMember] onEvent:^(MXEvent *event, MXTimelineDirection direction, MXRoomState *roomState) {
 
-            lastKnownMembership = room.state.membership;
+            lastKnownMembership = room.summary.membership;
         }];
         
         // This implicitly tests MXSession leaveRoom
@@ -211,13 +211,13 @@
 
                     MXRoom *room = [mxSession roomWithRoomId:roomId];
 
-                    XCTAssertEqual(room.state.membership, MXMembershipInvite);
-                    XCTAssertEqual(room.state.membersCount.members, 2, @"The room state information is limited while the room is joined");
+                    XCTAssertEqual(room.summary.membership, MXMembershipInvite);
+                    XCTAssertEqual(room.summary.membersCount.members, 2, @"The room state information is limited while the room is joined");
 
                     [room join:^{
 
-                        XCTAssertEqual(room.state.membership, MXMembershipJoin);
-                        XCTAssertEqual(room.state.membersCount.members, 2, @"The room state must be fully known (after an initialSync on the room");
+                        XCTAssertEqual(room.summary.membership, MXMembershipJoin);
+                        XCTAssertEqual(room.summary.membersCount.members, 2, @"The room state must be fully known (after an initialSync on the room");
 
                         [expectation fulfill];